+++ programs/drift/src/controller/repeg/tests.rs
@@ -8,6 +8,7 @@ use crate::math::repeg::{
     calculate_fee_pool, calculate_peg_from_target_price, calculate_repeg_cost,
 };
 use crate::state::oracle::HistoricalOracleData;
+use crate::state::oracle::OraclePriceData;
 use crate::state::perp_market::{ContractTier, AMM};
 use crate::state::state::{PriceDivergenceGuardRails, ValidityGuardRails};
 
@@ -51,8 +52,8 @@ pub fn update_amm_test() {
     let state = State {
         oracle_guard_rails: OracleGuardRails {
             price_divergence: PriceDivergenceGuardRails {
-                mark_oracle_divergence_numerator: 1,
-                mark_oracle_divergence_denominator: 10,
+                mark_oracle_percent_divergence: 1,
+                oracle_twap_5min_percent_divergence: 10,
             },
             validity: ValidityGuardRails {
                 slots_before_stale_for_amm: 10,     // 5s
@@ -193,8 +194,8 @@ pub fn update_amm_test_bad_oracle() {
     let state = State {
         oracle_guard_rails: OracleGuardRails {
             price_divergence: PriceDivergenceGuardRails {
-                mark_oracle_divergence_numerator: 1,
-                mark_oracle_divergence_denominator: 10,
+                mark_oracle_percent_divergence: 1,
+                oracle_twap_5min_percent_divergence: 10,
             },
             validity: ValidityGuardRails {
                 slots_before_stale_for_amm: 10,      // 5s
@@ -239,8 +240,8 @@ pub fn update_amm_larg_conf_test() {
     let state = State {
         oracle_guard_rails: OracleGuardRails {
             price_divergence: PriceDivergenceGuardRails {
-                mark_oracle_divergence_numerator: 1,
-                mark_oracle_divergence_denominator: 10,
+                mark_oracle_percent_divergence: 1,
+                oracle_twap_5min_percent_divergence: 10,
             },
             validity: ValidityGuardRails {
                 slots_before_stale_for_amm: 10,      // 5s
@@ -370,8 +371,8 @@ pub fn update_amm_larg_conf_w_neg_tfmd_test() {
     let state = State {
         oracle_guard_rails: OracleGuardRails {
             price_divergence: PriceDivergenceGuardRails {
-                mark_oracle_divergence_numerator: 1,
-                mark_oracle_divergence_denominator: 10,
+                mark_oracle_percent_divergence: 1,
+                oracle_twap_5min_percent_divergence: 10,
             },
             validity: ValidityGuardRails {
                 slots_before_stale_for_amm: 10,      // 5s