+++ programs/drift/src/math/margin.rs
@@ -16,18 +16,16 @@ use crate::math::casting::Cast;
 use crate::math::funding::calculate_funding_payment;
 use crate::math::oracle::{is_oracle_valid_for_action, DriftAction};
 
-use crate::math::spot_balance::{
-    get_balance_value_and_token_amount, get_strict_token_value, get_token_value,
-};
+use crate::math::spot_balance::{get_strict_token_value, get_token_value};
 
 use crate::math::safe_math::SafeMath;
 use crate::state::oracle::OraclePriceData;
 use crate::state::oracle_map::OracleMap;
 use crate::state::perp_market::{ContractTier, MarketStatus, PerpMarket};
 use crate::state::perp_market_map::PerpMarketMap;
-use crate::state::spot_market::{AssetTier, SpotBalanceType, SpotMarket};
+use crate::state::spot_market::{AssetTier, SpotBalanceType};
 use crate::state::spot_market_map::SpotMarketMap;
-use crate::state::user::{PerpPosition, SpotPosition, User};
+use crate::state::user::{PerpPosition, User};
 use num_integer::Roots;
 use solana_program::msg;
 use std::cmp::{max, min, Ordering};
@@ -39,6 +37,7 @@ mod tests;
 #[derive(Clone, Copy, PartialEq, Debug, Eq)]
 pub enum MarginRequirementType {
     Initial,
+    Fill,
     Maintenance,
 }
 
@@ -97,35 +96,6 @@ pub fn calculate_size_discount_asset_weight(
     Ok(min_asset_weight)
 }
 
-pub fn calculate_spot_position_value(
-    spot_position: &SpotPosition,
-    spot_market: &SpotMarket,
-    oracle_price_data: &OraclePriceData,
-    margin_requirement_type: MarginRequirementType,
-) -> DriftResult<u128> {
-    let (balance_value, token_amount) =
-        get_balance_value_and_token_amount(spot_position, spot_market, oracle_price_data)?;
-
-    let balance_equity_value = match spot_position.balance_type {
-        SpotBalanceType::Deposit => balance_value
-            .safe_mul(
-                spot_market
-                    .get_asset_weight(token_amount, &margin_requirement_type)?
-                    .cast()?,
-            )?
-            .safe_div(SPOT_WEIGHT_PRECISION_U128)?,
-        SpotBalanceType::Borrow => balance_value
-            .safe_mul(
-                spot_market
-                    .get_liability_weight(token_amount, &margin_requirement_type)?
-                    .cast()?,
-            )?
-            .safe_div(SPOT_WEIGHT_PRECISION_U128)?,
-    };
-
-    Ok(balance_equity_value)
-}
-
 pub fn calculate_perp_position_value_and_pnl(
     market_position: &PerpPosition,
     market: &PerpMarket,
@@ -143,10 +113,22 @@ pub fn calculate_perp_position_value_and_pnl(
         oracle_price_data.price
     };
 
+    // the funding must be calculated before calculated the unrealized pnl w simulated lp position
+    let unrealized_funding = calculate_funding_payment(
+        if market_position.base_asset_amount > 0 {
+            market.amm.cumulative_funding_rate_long
+        } else {
+            market.amm.cumulative_funding_rate_short
+        },
+        market_position,
+    )?;
+
     let market_position = market_position.simulate_settled_lp_position(market, valuation_price)?;
 
-    let total_unrealized_pnl =
-        calculate_total_unrealized_perp_pnl(&market_position, market, valuation_price)?;
+    let (_, unrealized_pnl) =
+        calculate_base_asset_value_and_pnl_with_oracle_price(&market_position, valuation_price)?;
+
+    let total_unrealized_pnl = unrealized_pnl.safe_add(unrealized_funding.cast()?)?;
 
     let worst_case_base_asset_amount = market_position.worst_case_base_asset_amount()?;
 
@@ -210,26 +192,6 @@ pub fn calculate_perp_position_value_and_pnl(
         weighted_unrealized_pnl,
         worse_case_base_asset_value,
     ))
-}
-
-pub fn calculate_total_unrealized_perp_pnl(
-    market_position: &PerpPosition,
-    market: &PerpMarket,
-    valuation_price: i64,
-) -> DriftResult<i128> {
-    let unrealized_funding = calculate_funding_payment(
-        if market_position.base_asset_amount > 0 {
-            market.amm.cumulative_funding_rate_long
-        } else {
-            market.amm.cumulative_funding_rate_short
-        },
-        market_position,
-    )?;
-
-    let (_, unrealized_pnl) =
-        calculate_base_asset_value_and_pnl_with_oracle_price(market_position, valuation_price)?;
-
-    unrealized_pnl.safe_add(unrealized_funding.cast()?)
 }
 
 pub fn calculate_user_safest_position_tiers(
@@ -663,7 +625,7 @@ pub fn meets_withdraw_margin_requirement(
         strict,
     )?;
 
-    if initial_margin_requirement > 0 {
+    if initial_margin_requirement > 0 || num_of_liabilities > 0 {
         validate!(
             oracles_valid,
             ErrorCode::InvalidOracle,
@@ -879,10 +841,12 @@ pub fn validate_spot_margin_trading(
         }
     }
 
-    let quote_spot_market = spot_market_map.get_quote_spot_market()?;
-    let quote_token_amount = user
-        .get_quote_spot_position()
-        .get_signed_token_amount(&quote_spot_market)?;
+    let mut quote_token_amount = 0_i128;
+    let quote_spot_position = user.get_quote_spot_position();
+    if !quote_spot_position.is_available() {
+        let quote_spot_market = spot_market_map.get_quote_spot_market()?;
+        quote_token_amount = quote_spot_position.get_signed_token_amount(&quote_spot_market)?;
+    }
 
     // The user can have open bids if their value is less than existing quote token amount
     validate!(
@@ -960,10 +924,24 @@ pub fn calculate_user_equity(
             oracle_price_data.price
         };
 
+        let unrealized_funding = calculate_funding_payment(
+            if market_position.base_asset_amount > 0 {
+                market.amm.cumulative_funding_rate_long
+            } else {
+                market.amm.cumulative_funding_rate_short
+            },
+            market_position,
+        )?;
+
         let market_position =
             market_position.simulate_settled_lp_position(market, valuation_price)?;
 
-        let pnl = calculate_total_unrealized_perp_pnl(&market_position, market, valuation_price)?;
+        let (_, unrealized_pnl) = calculate_base_asset_value_and_pnl_with_oracle_price(
+            &market_position,
+            valuation_price,
+        )?;
+
+        let pnl = unrealized_pnl.safe_add(unrealized_funding.cast()?)?;
 
         let pnl_value = pnl
             .safe_mul(quote_oracle_price.cast()?)?