+++ sdk/src/constants/perpMarkets.ts
@@ -121,7 +121,7 @@ export const DevnetPerpMarkets: PerpMarketConfig[] = [
 		baseAssetSymbol: '1MPEPE',
 		marketIndex: 10,
 		oracle: new PublicKey('Gz9RfgDeAFSsH7BHDGyNTgCik74rjNwsodJpsCizzmkj'),
-		launchTs: 1683125906000,
+		launchTs: 1683781239000,
 		oracleSource: OracleSource.PYTH_1M,
 	},
 	{
@@ -131,7 +131,7 @@ export const DevnetPerpMarkets: PerpMarketConfig[] = [
 		baseAssetSymbol: 'OP',
 		marketIndex: 11,
 		oracle: new PublicKey('8ctSiDhA7eJoii4TkKV8Rx4KFdz3ycsA1FXy9wq56quG'),
-		launchTs: 1683125906000,
+		launchTs: 1686091480000,
 		oracleSource: OracleSource.PYTH,
 	},
 	{
@@ -141,7 +141,7 @@ export const DevnetPerpMarkets: PerpMarketConfig[] = [
 		baseAssetSymbol: 'RNDR',
 		marketIndex: 12,
 		oracle: new PublicKey('C2QvUPBiU3fViSyqA4nZgGyYqLgYf9PRpd8B8oLoo48w'),
-		launchTs: 1683125906000,
+		launchTs: 1687201081000,
 		oracleSource: OracleSource.PYTH,
 	},
 ];
@@ -254,7 +254,7 @@ export const MainnetPerpMarkets: PerpMarketConfig[] = [
 		baseAssetSymbol: '1MPEPE',
 		marketIndex: 10,
 		oracle: new PublicKey('FSfxunDmjjbDV2QxpyxFCAPKmYJHSLnLuvQXDLkMzLBm'),
-		launchTs: 1683125906000,
+		launchTs: 1683781239000,
 		oracleSource: OracleSource.PYTH_1M,
 	},
 	{
@@ -264,7 +264,7 @@ export const MainnetPerpMarkets: PerpMarketConfig[] = [
 		baseAssetSymbol: 'OP',
 		marketIndex: 11,
 		oracle: new PublicKey('4o4CUwzFwLqCvmA5x1G4VzoZkAhAcbiuiYyjWX1CVbY2'),
-		launchTs: 1683125906000,
+		launchTs: 1686091480000,
 		oracleSource: OracleSource.PYTH,
 	},
 	{
@@ -274,7 +274,7 @@ export const MainnetPerpMarkets: PerpMarketConfig[] = [
 		baseAssetSymbol: 'RNDR',
 		marketIndex: 12,
 		oracle: new PublicKey('CYGfrBJB9HgLf9iZyN4aH5HvUAi2htQ4MjPxeXMf4Egn'),
-		launchTs: 1683125906000,
+		launchTs: 1687201081000,
 		oracleSource: OracleSource.PYTH,
 	},
 ];