+++ sdk/tests/dlob/helpers.ts
@@ -262,6 +262,8 @@ export const mockSpotMarkets: Array<SpotMarketAccount> = [
 		pubkey: PublicKey.default,
 		mint: DevnetSpotMarkets[0].mint,
 		vault: PublicKey.default,
+		minOrderSize: ZERO,
+		maxPositionSize: ZERO,
 		revenuePool: {
 			scaledBalance: new BN(0),
 			marketIndex: 0,
@@ -344,6 +346,8 @@ export const mockSpotMarkets: Array<SpotMarketAccount> = [
 			scaledBalance: new BN(0),
 			marketIndex: 0,
 		},
+		minOrderSize: ZERO,
+		maxPositionSize: ZERO,
 		insuranceFund: {
 			vault: PublicKey.default,
 			totalShares: new BN(0),
@@ -422,6 +426,8 @@ export const mockSpotMarkets: Array<SpotMarketAccount> = [
 			scaledBalance: new BN(0),
 			marketIndex: 0,
 		},
+		minOrderSize: ZERO,
+		maxPositionSize: ZERO,
 		insuranceFund: {
 			vault: PublicKey.default,
 			totalShares: new BN(0),
@@ -506,8 +512,8 @@ export const mockStateAccount: StateAccount = {
 	liquidationDuration: 0,
 	oracleGuardRails: {
 		priceDivergence: {
-			markOracleDivergenceNumerator: new BN(0),
-			markOracleDivergenceDenominator: new BN(0),
+			markOraclePercentDivergence: new BN(0),
+			oracleTwap5MinPercentDivergence: new BN(0),
 		},
 		validity: {
 			slotsBeforeStaleForAmm: new BN(0),